+++ ros/wavemap_ros/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package wavemap_ros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (2023-08-09)
+------------------
+* Enable file saving in ROS server
+* Contributors: Victor Reijgwart
+
 1.0.0 (2023-08-08)
 ------------------
 * First public release