+++ ros/wavemap_ros_conversions/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package wavemap_ros_conversions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (2023-08-09)
+------------------
+
 1.0.0 (2023-08-08)
 ------------------
 * First public release